requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

sqlalchemy
psycopg2-binary

chromadb

streamlit

unstructured
langchain
langchain_core
langchain_community
langchain-chroma
langchain_openai
langchain[google-genai]
langgraph
langchain-google-genai
google-genai
sentence-transformers

# python-magic # for Linux/Mac
python-magic-bin